refactor(ui): extract perform pad rendering
