+++ crates/vibez-ui/src/app/mod.rs
@@ -159,6 +159,7 @@ mod views_devices;
 mod views_mixer;
 mod views_overlays;
 mod views_perform;
+mod views_perform_pads;
 mod views_perform_playhead;
 mod views_perform_sections;
 mod views_settings;
